Fixed a bug in the Evaluator class by converting map object to list.
Fixed bug in the iou metric by specifying in a one hot operation on the predictions that the number of classes must be equal to predictions.shape[1]. Otherwise if the highest numbered class was never predicted as most likely, then the argmax operation would return only lower numbers and one_hot would then underestimate the number of classes, since by default it assumes the highest number in the tensor it is given equals the number of classes.

Changed the behaviour of RasterioDataset to 'warnings.warn([MESSAGE], MaskNotPresentWarning)' instead of 'raise MaskNotPresentWarning' as this interrupted the execution of the program.