Added an ability to get load rasters from paths provided by a custom text file format.
Changed the criteria for the masks so that the date of a tornado must match the stormdate on the raster for that tornado to be used as a mask for the raster.

dded getters to Evaluator.

Removed extraneous imports in main.py and changed the loss function to use only the IoU for the positive class (tornado present at pixel).

Modified the RasterioDataset and training DataLoader, changed batch size to 1 (since images are now not of a standardized size), and added graphing of the results.

Changed the neural net to have 3 layers, use a 3x3 kernel size, have 64 channels in the intermediate layers, and use batch normalization between intermediate layers. Increased the learning rate by 100x from 1e-5 to 1e-3.

Modified rasterio dataset to get rasters from a loader function provided as a parameter instead of retrieving them itself by recursively walking through a provided path.

Added a get_rasters_from_file function in utils.py that retrieves a list of directories from a specially formatted text file and loads rasters from them.